+++ src/SOAPify/analysis.py
@@ -2,19 +2,23 @@
 
 import numpy
 from numpy import ndarray
-from .distances import simpleSOAPdistance
 from MDAnalysis import Universe, AtomGroup
 from MDAnalysis.lib.NeighborSearch import AtomNeighborSearch
+import h5py
+
+from .distances import simpleSOAPdistance
+from .utils import getSOAPSettings, normalizeArray, fillSOAPVectorFromdscribe
 
 
-def tempoSOAP(
+def timeSOAP(
     SOAPTrajectory: ndarray,
     window: int = 1,
     stride: int = None,
     backward: bool = False,
+    returnDiff: bool = True,
     distanceFunction: callable = simpleSOAPdistance,
 ) -> "tuple[ndarray, ndarray]":
-    """performs the 'tempoSOAP' analysis on the given SOAP trajectory
+    """performs the 'timeSOAP' analysis on the given SOAP trajectory
 
         * Original author: Cristina Caruso
         * Mantainer: Daniele Rapetti
@@ -27,14 +31,18 @@ def tempoSOAP(
         stride (int):
             the stride in frames between each state confrontation. **NOT IN USE**.
             Defaults to None.
-        deltaT (int): number of frames to skip
+        backward (bool):
+            If true the soap distance is referred to the previous frame.
+             **NOT IN USE**. Defaulst to True.
+        returnDiff (bool):
+            If true returns also the first derivative of timeSOAP. Defaults to True.
         distanceFunction (callable, optional):
             the function that define the distance. Defaults to :func:`SOAPify.distances.simpleSOAPdistance`.
 
     Returns:
         tuple[numpy.ndarray,numpy.ndarray]:
-            - **timedSOAP** the tempoSOAP values, shape(frames-1,natoms)
-            - **deltaTimedSOAP** the derivatives of tempoSOAP, shape(natoms, frames-2)
+            - **timedSOAP** the timeSOAP values, shape(frames-1,natoms)
+            - **deltaTimedSOAP** the derivatives of timeSOAP, shape(natoms, frames-2)
     """
     if stride is None:
         stride = window
@@ -55,18 +63,20 @@ def tempoSOAP(
     # )
     # print(SOAPTrajectory.shape)
 
-    deltaTimedSOAP = numpy.diff(timedSOAP.T, axis=-1)
+    if returnDiff:
+        deltaTimedSOAP = numpy.diff(timedSOAP.T, axis=-1)
+        return timedSOAP, deltaTimedSOAP
+    return timedSOAP
 
-    return timedSOAP, deltaTimedSOAP
 
-
-def tempoSOAPsimple(
+def timeSOAPsimple(
     SOAPTrajectory: ndarray,
     window: int = 1,
     stride: int = None,
     backward: bool = False,
+    returnDiff: bool = True,
 ) -> "tuple[ndarray, ndarray]":
-    r"""performs the 'tempoSOAP' analysis on the given **normalized** SOAP trajectory
+    r"""performs the 'timeSOAP' analysis on the given **normalized** SOAP trajectory
 
         this is optimized to use :func:`SOAPify.distances.simpleSOAPdistance`,
         without calling it.
@@ -100,12 +110,16 @@ def tempoSOAPsimple(
         stride (int):
             the stride in frames between each state confrontation. **NOT IN USE**.
             Defaults to None.
-        deltaT (int): number of frames to skip
+        backward (bool):
+            If true the soap distance is referred to the previous frame.
+             **NOT IN USE**. Defaulst to True.
+        returnDiff (bool):
+            If true returns also the first derivative of timeSOAP. Defaults to True.
 
     Returns:
         tuple[numpy.ndarray,numpy.ndarray]:
-            - **timedSOAP** the tempoSOAP values, shape(frames-1,natoms)
-            - **deltaTimedSOAP** the derivatives of tempoSOAP, shape(natoms, frames-2)
+            - **timedSOAP** the timeSOAP values, shape(frames-1,natoms)
+            - **deltaTimedSOAP** the derivatives of timeSOAP, shape(natoms, frames-2)
     """
     if stride is None:
         stride = window
@@ -122,9 +136,71 @@ def tempoSOAPsimple(
         timedSOAP[frame - window] = numpy.linalg.norm(actual - prev, axis=1)
         prev = actual
 
-    deltaTimedSOAP = numpy.diff(timedSOAP.T, axis=-1)
+    if returnDiff:
+        deltaTimedSOAP = numpy.diff(timedSOAP.T, axis=-1)
+        return timedSOAP, deltaTimedSOAP
+    return timedSOAP
 
-    return timedSOAP, deltaTimedSOAP
+
+def getTimeSOAPSimple(
+    soapDataset: h5py.Dataset,
+    window: int = 1,
+    stride: int = None,
+    backward: bool = False,
+):
+    """Shortcut to extract the timeSOAP from large datasets.
+
+        This function is the equivalent to:
+
+        - loading a chunk of the trajectory from a h5py.Dataset with a SOAP fingerprints trajectory
+        - filling the vector with :func:`SOAPify.utils.fillSOAPVectorFromdscribe`
+        - normalizing it with :func:`SOAPify.utils.normalizeArray`
+        - calculating the timeSOAP with  :func:`timeSOAPsimple`
+        and then returning timeSOAP and the derivative
+
+
+    Args:
+        soapDataset (h5py.Dataset):
+            the dataset with the SOAP fingerprints
+        window (int):
+            the dimension of the windows between each state confrontations.
+            See :func:`timeSOAPsimple`
+            Defaults to 1.
+        stride (int):
+            the stride in frames between each state confrontation.
+            See :func:`timeSOAPsimple`
+            Defaults to None.
+        backward (bool):
+            If true the soap distance is referred to the previous frame.
+            See :func:`timeSOAPsimple` . Defaulst to True.
+
+    Returns:
+        tuple[numpy.ndarray,numpy.ndarray]:
+            - **timedSOAP** the timeSOAP values, shape(frames-1,natoms)
+            - **deltaTimedSOAP** the derivatives of timeSOAP, shape(natoms, frames-2)
+    """
+    fillSettings = getSOAPSettings(soapDataset)
+    timedSOAP = numpy.zeros((soapDataset.shape[0] - window, soapDataset.shape[1]))
+    # TODO: add a check to the window
+
+    slide = 0
+    # this looks a lot convoluted, but it is way faster than working one atom
+    # at a time
+    for c in soapDataset.iter_chunks():
+        theSlice = slice(c[0].start - slide, c[0].stop, c[0].step)
+        outSlice = slice(c[0].start - slide, c[0].stop - 1, c[0].step)
+        timedSOAP[outSlice] = timeSOAPsimple(
+            normalizeArray(
+                fillSOAPVectorFromdscribe(soapDataset[theSlice], **fillSettings)
+            ),
+            window=window,
+            stride=stride,
+            backward=backward,
+            returnDiff=False,
+        )
+        slide = 1
+
+    return timedSOAP, numpy.diff(timedSOAP.T, axis=-1)
 
 
 def listNeighboursAlongTrajectory(